Fixed: 1.Changed the root file dictionary to edu.usf.sas.pal.muser 2.Added a .gitignore file 3.Updated all copyright headers to be 2019-2020 4.Auto-formated all files and applied the IntelliJ default template. 5.Renamed the class varibles from tbr1 to mam1 6.Removed m prefix from all the class-level variables
