+++ muser/src/main/java/edu/usf/cutr/mfe/exception/FirebaseFileNotInitializedException.java
@@ -1,0 +1,7 @@
+package edu.usf.cutr.mfe.exception;
+
+public class FirebaseFileNotInitializedException extends Exception {
+	public FirebaseFileNotInitializedException() {
+		super("Firebase file is not initialized.");
+	}
+}
