+++ muser/src/main/java/edu/usf/cutr/mfe/utils/StringUtils.java
@@ -1,0 +1,8 @@
+package edu.usf.cutr.mfe.utils;
+public class StringUtils {
+	public static String valueOf(Object o) {
+		if (o == null)
+			return "";
+		return String.valueOf(o);
+	}
+}
