--- muser/src/main/java/edu/usf/sas/pal/muser/constants/FirebaseConstants.java
@@ -1,13 +1,0 @@
-package edu.usf.sas.pal.muser.constants;
-
-public class FirebaseConstants {
-
-    public static final String CREDENTIAL_FILE = "admin-key.json";
-
-    public static final String MUSIC_ANALYSIS_CSV_FILE = "music-analysis.csv";
-
-    public static final String FIREBASE_PLAYER_EVENTS_FOLDER = "player-events";
-
-    public static final String FIREBASE_UI_EVENTS_FOLDER = "ui-events";
-
-}
