--- muser/src/main/java/edu/usf/sas/pal/muser/exception/FirebaseFileNotInitializedException.java
@@ -1,7 +1,0 @@
-package edu.usf.sas.pal.muser.exception;
-
-public class FirebaseFileNotInitializedException extends Exception {
-    public FirebaseFileNotInitializedException() {
-        super("Firebase file is not initialized.");
-    }
-}
