--- muser/src/main/java/edu/usf/sas/pal/muser/utils/StringUtils.java
@@ -1,9 +1,0 @@
-package edu.usf.sas.pal.muser.utils;
-
-public class StringUtils {
-    public static String valueOf(Object o) {
-        if (o == null)
-            return "";
-        return String.valueOf(o);
-    }
-}
