+++ muser/src/main/resources/.gitignore
@@ -117,6 +117,7 @@ hs_err_pid*
 .idea/**/gradle.xml
 .idea/**/libraries
 
+
 # Gradle and Maven with auto-import
 # When using Gradle or Maven with auto-import, you should exclude module files,
 # since they will be recreated, and may cause churn.  Uncomment if using
